06_a2a_communication/pyproject.toml
[project]
name = "06-agent-debate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "openai-agents[litellm]>=0.3.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
    "uvicorn>=0.35.0",
]
